ark/syntropy-ark-bot/requirements.txt
aiohttp==3.7.4
aioify==0.4.0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
chardet==3.0.4
discord.py==1.7.3
docker==5.0.3
idna==3.2
idna-ssl==1.1.0
module-wrapper==0.3.1
multidict==5.2.0
poetry-version==0.1.5
python-dateutil==2.8.2
python-dotenv==0.19.1
requests==2.26.0
six==1.16.0
syntropy-sdk==0.3.0
tomlkit==0.5.11
typing-extensions==3.10.0.2
urllib3==1.26.7
websocket-client==1.2.1
yarl==1.7.0
loguru==0.5.3